fix: use extra storage class when reading secrets
When secrets are read from annotations automatically then the storage
class will not support creating a PVC without the annotations. Therefore
we add the option to get the secret from the PVC annotations from a new
storage class so that when the old storage class is used things work as
before.